requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==2.0
django-rest-coreapi-schema==0.1.2
django-rest-swagger==2.2.0
djangorestframework==3.7.7
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
jsonhyperschema-codec==1.0.3
Markdown==2.6.10
MarkupSafe==1.1.1
openapi-codec==1.3.2
pkg-resources==0.0.0
Pygments==2.2.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
simplejson==3.17.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
